Adds support for custom Hex repositories
Allows the deployment process to add custom Hex repositories,
specifically enabling the use of the petal.build repository for
dependency resolution.

This involves adding a mix command to add the custom repository with
authentication, and fetching dependencies only for the production
environment.  It also includes vault integration for sensitive information.